requirements.txt
streamlit==1.37.0
python-dotenv